refactor: move PDP sample data to __mocks__
