+++ src/components/ProductDetail/ProductDetail.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import renderer from "react-test-renderer";
 import { shallow } from "enzyme";
 import ProductDetail from "./ProductDetail";
-import sampleData from "./sampleData";
+import sampleData from "./__mocks__/sampleData";
 
 test("basic snapshot", () => {
   const component = renderer.create(shallow(<ProductDetail catalogProduct={sampleData} />).get(0));
