+++ src/components/ProductDetail/__mocks__/sampleData.js
